# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Mandaean Flutter App API",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/posts": {
            "get": {
                "tags": [
                    "Community"
                ],
                "summary": "Get all posts",
                "operationId": "3fda362e2e082b5357ac416e7767e435",
                "responses": {
                    "200": {
                        "description": "List of posts with user, likes, comments, shares."
                    }
                },
                "security": [
                    {
                        "apiKey": []
                    },
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Community"
                ],
                "summary": "Create a new post",
                "operationId": "dcb66200c3132996151884ac232cf439",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "content"
                                ],
                                "properties": {
                                    "content": {
                                        "type": "string",
                                        "example": "My story..."
                                    },
                                    "image": {
                                        "description": "Image file to upload",
                                        "type": "file"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Post created."
                    }
                },
                "security": [
                    {
                        "apiKey": []
                    },
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/posts/{post}": {
            "get": {
                "tags": [
                    "Community"
                ],
                "summary": "Get a single post with details",
                "operationId": "5042caf9a6c47bde39921c8f400e28b5",
                "parameters": [
                    {
                        "name": "post",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Post details."
                    }
                },
                "security": [
                    {
                        "apiKey": []
                    },
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Community"
                ],
                "summary": "Update a post",
                "operationId": "fc17b67462ad71a7f4b287520462fd02",
                "parameters": [
                    {
                        "name": "post",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "content"
                                ],
                                "properties": {
                                    "content": {
                                        "type": "string",
                                        "example": "Updated story..."
                                    },
                                    "image": {
                                        "description": "Image file to upload (optional)",
                                        "type": "file"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Post updated."
                    }
                },
                "security": [
                    {
                        "apiKey": []
                    },
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Community"
                ],
                "summary": "Delete a post",
                "operationId": "f0dcb391948b9589ee732a26a68e17c7",
                "parameters": [
                    {
                        "name": "post",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Post deleted."
                    }
                },
                "security": [
                    {
                        "apiKey": []
                    },
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/posts/{post}/like": {
            "post": {
                "tags": [
                    "Community"
                ],
                "summary": "Like or unlike a post",
                "operationId": "480c2fcc9b06d93a9061fab6af539b4d",
                "parameters": [
                    {
                        "name": "post",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Like toggled."
                    }
                },
                "security": [
                    {
                        "apiKey": []
                    },
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/posts/{post}/comment": {
            "post": {
                "tags": [
                    "Community"
                ],
                "summary": "Add a comment to a post",
                "operationId": "e77cd6b573611a1fdded552c317b27e8",
                "parameters": [
                    {
                        "name": "post",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "content"
                                ],
                                "properties": {
                                    "content": {
                                        "type": "string",
                                        "example": "Nice post!"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Comment added."
                    }
                },
                "security": [
                    {
                        "apiKey": []
                    },
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/posts/{post}/share": {
            "post": {
                "tags": [
                    "Community"
                ],
                "summary": "Share a post",
                "operationId": "31ce760abd8957275c8b5bc6b5ea13f7",
                "parameters": [
                    {
                        "name": "post",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Post shared."
                    }
                },
                "security": [
                    {
                        "apiKey": []
                    },
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/my-posts": {
            "get": {
                "tags": [
                    "Community"
                ],
                "summary": "Get posts of the authenticated user",
                "operationId": "b1b8c2c6c33c80f172a3c5694b386acd",
                "responses": {
                    "200": {
                        "description": "List of user's own posts."
                    }
                },
                "security": [
                    {
                        "apiKey": []
                    },
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            },
            "apiKey": {
                "type": "apiKey",
                "name": "X-API-KEY",
                "in": "header"
            }
        }
    },
...
}